+++ src/ContentElements/Person.php
@@ -35,13 +35,16 @@ class Person extends \ContentElement
 				}
 
 				// Links bauen
-				$links = unserialize($objPerson->links);
+				$links = (array)unserialize($objPerson->links);
 				$linkdaten = '';
-				foreach($links as $item)
+				if(is_array($links) && count($links) > 0)
 				{
-					if($item['active'])
+					foreach($links as $item)
 					{
-						$linkdaten .= $item['url'] ? '[<a href="'.$item['url'].'"'.($item['target'] ? ' target="_blank"' : '').'>'.($item['text'] ? $item['text'] : $item['url']).'</a>] ' : $item['text'];
+						if(isset($item['active']) && $item['active'])
+						{
+							$linkdaten .= $item['url'] ? '[<a href="'.$item['url'].'"'.($item['target'] ? ' target="_blank"' : '').'>'.($item['text'] ? $item['text'] : $item['url']).'</a>] ' : $item['text'];
+						}
 					}
 				}